Replace antelope-controlplane + master-watcher jobs by antelope + epoxy
master branch is not longer supported in centos9 (python3.9) upstream,
so we should avoid using it specially in a stable job with antelope
based controlplane.

This patch is replacing the jobs using watcher from master in antelope
controlplane jobs by epoxy based watcher containers. The resulting
testing combination is:

1. Antelope controlplane + epoxy watcher (centos 9)
- Job on ocp 4.18
- Job on ocp 4.16

2. Master controlplane + master watcher (centos 9)
- Job on ocp 4.18

Note we will need to migrate [2] to centos10 which is work in progress
right now.