Remove route creation from the watcher-operator
This is part of the integration in the OpenStack Operator.

Once integrated on it, the route will be managed by it so we need to get
rid of all the logic existing in the operator for it and the apiOverride
parameter which was intended to pass all the parameters required for the
Watcher controller to manage it.

This patch is also adjusting functional and kuttl tests and ci
deployment files for it.

Jira: OSPRH-16238