Move parameters to pointers
When integrating the watcher-operator in the openstack-operator, the
defaulting webhooks can interfere with the kubebuilder defaults.

The defaulting webhook can set fields to their zero-values.  In the case
of a string, then "".  This happens simply because of the serialization
that occurs in the webhooks.  Then when the field arrives in the kubebuilder
defaulting logic, it has a non-null value, "", so the kubebuilder leaves it
alone and it then eventually arrives in the reconcile loop as "".

To avoid that, I'm moving to use pointers to all string parameters which
has a kubebuilder default value.