Configure min and max microversion for watcher-tempest-plugin
A new upstream patch is adding microversion testing for
watcher-tempest-plugin, and it will be mandatory to configure
the min and max microversion to be tested in each job.
To make the opendev job to work with new microversions, we can
set max_microversion to 'latest' and for stable branches jobs we
override with the latest microversion for each release.
This PR also sets 1.4 as max microversion for Epoxy.
It also renames the test configuration file to replaces the suffix
master since this file is being used in all jobs.

Depends-On: https://review.opendev.org/c/openstack/watcher-tempest-plugin/+/956306
Signed-off-by: Douglas Viroel <viroel@gmail.com>