Add support for consuming notifications
This patch is adding a new parameter `notificationsBusInstance` to the
Watcher CRD. The value should be the name of a rabbitmq instance which
is created specifically for notifications, not RPCs.

When it has a value, the watcher-operator will create a TransportURL
object for credentials and will configure it as transport_url in the
`oslo_messaging_notifications` section so that the dedision engine and
potentially any watcher service can consume Nova or other services
notifications.

Note that, althoug watcher can also send notifications, this patch is
disabling sending notifications by default as there is no known
consumers of those messages.

The patch is also adding functional and kuttl tests.

Jira: OSPRH-18694

Signed-off-by: Alfredo Moralejo <amoralej@redhat.com>