Bump watcher-tempest-plugin for epoxy jobs to known good commit
Currently, watcher-tempest-plugin is pinned in RDO to tag 3.4.0 [1].

Some fixes and improvements have been merged since then and the required
backports in the watcher epoxy branch are merged, so let's bump the
tempest plugin.

Given that in the past it has been usual that patches in master branch
of the plugin requires backports in the watcher epoxy branch, I am not
proposing to follow the master branch of the plugin but using a known
good commit.

I'm disabling one zone_migration test which requires [2] to be merged.

[1] https://github.com/redhat-openstack/rdoinfo/blob/ee923461f857154321314796ed06c8365c1457d0/tags/epoxy.yml#L943-L946
[2] https://review.opendev.org/c/openstack/watcher/+/951535

Signed-off-by: Alfredo Moralejo <amoralej@redhat.com>