Add delay between retries when waiting for prometheus pod
In some cases, the job executes the oc wait right in the moment when the
pod has been deleted and the new one is not created yet. In that
situation oc wait fails with a `pods ... not found` error and exits
immediately, so the 1 minute timeout is not applied and the 10 retries
are executed immediately.

To avoid that, I'm adding a delay between retries so that, it waits for
before re-executing it so that it has time for the pod to be created.