Set image_base in cifmw_operator_build_operators for meta operator
image_base is used to specify the operator project name. It is used
openstack operator to include watcher operator catalog image into
the openstack-operator catalog image.

Sometimes image_base is set to empty when there is no github operator
depends-on. This pr sets the same in cifmw_operator_build_operators var.

The same is already added in telemetry-operator[1].

Links:
[1]. https://github.com/openstack-k8s-operators/telemetry-operator/pull/724

Signed-off-by: Chandan Kumar (raukadah) <raukadah@gmail.com>