Set zuul_log_collection to false to trigger must-gather facts #251
Before moving to role, when the 99-logs.yml playbook was executed, the
variables was not set to, so it pick default value which is false, so
the playbook was executed. Now when it is in `cifmw_setup` role in
`run_logs.yml` task file, the condition is possitive, so it is skipped.

Add 'zuul_log_collection' var to overwrite other existing vars and
trigger gather facts.

Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/3208
Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/3231
Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>